# api-gateway/docs/swagger.yaml
basePath: /
definitions:
  api-gateway_internal_models.ChangePasswordRequest:
    properties:
      new_password:
        example: newpassword123
        minLength: 6
        type: string
      old_password:
        example: oldpassword123
        type: string
    required:
    - new_password
    - old_password
    type: object
  api-gateway_internal_models.ErrorResponse:
    properties:
      error:
        example: Error message
        type: string
    type: object
  api-gateway_internal_models.ForgotPasswordRequest:
    properties:
      email:
        example: user@example.com
        type: string
    required:
    - email
    type: object
  api-gateway_internal_models.LoginRequest:
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: Password123!
        type: string
    required:
    - email
    - password
    type: object
  api-gateway_internal_models.LoginResponse:
    properties:
      access_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      message:
        example: Login successful
        type: string
      refresh_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      user:
        $ref: '#/definitions/api-gateway_internal_models.User'
    type: object
  api-gateway_internal_models.RegisterRequest:
    properties:
      email:
        example: user@example.com
        type: string
      first_name:
        example: John
        type: string
      last_name:
        example: Doe
        type: string
      password:
        example: Password123!
        type: string
      phone_number:
        example: "+1234567890"
        type: string
      push_notification_permission:
        example: true
        type: boolean
      role:
        example: user
        type: string
    required:
    - email
    - first_name
    - last_name
    - password
    type: object
  api-gateway_internal_models.RegisterResponse:
    properties:
      message:
        example: User created successfully
        type: string
      user:
        $ref: '#/definitions/api-gateway_internal_models.User'
    type: object
  api-gateway_internal_models.ResetPasswordRequest:
    properties:
      email:
        example: user@example.com
        type: string
      new_password:
        example: newpassword123
        minLength: 6
        type: string
      token:
        example: reset_token_here
        type: string
    required:
    - email
    - new_password
    - token
    type: object
  api-gateway_internal_models.SuccessResponse:
    properties:
      message:
        example: Operation successful
        type: string
      success:
        example: true
        type: boolean
    type: object
  api-gateway_internal_models.UpdateUserRequest:
    properties:
      email:
        example: newemail@example.com
        type: string
      first_name:
        example: Jane
        type: string
      last_name:
        example: Smith
        type: string
      phone_number:
        example: "+1234567890"
        type: string
      push_notification_permission:
        example: true
        type: boolean
      role:
        example: manager
        type: string
    type: object
  api-gateway_internal_models.User:
    properties:
      created_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      email:
        example: user@example.com
        type: string
      first_name:
        example: John
        type: string
      id:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
      last_name:
        example: Doe
        type: string
      phone_number:
        example: "+1234567890"
        type: string
      push_notification_permission:
        example: true
        type: boolean
      role:
        example: user
        type: string
    type: object
  api-gateway_internal_models.UserResponse:
    properties:
      message:
        example: User retrieved successfully
        type: string
      user:
        $ref: '#/definitions/api-gateway_internal_models.User'
    type: object
host: localhost:8080
info:
  contact: {}
  description: API Gateway for QAMQOR-VISION microservices
  title: QAMQOR-VISION API Gateway
  version: "1.0"
paths:
  /api/v1/forgot-password:
    post:
      consumes:
      - application/json
      description: Request password reset email
      parameters:
      - description: Email address
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api-gateway_internal_models.ForgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api-gateway_internal_models.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api-gateway_internal_models.ErrorResponse'
      summary: Forgot Password
      tags:
      - Authentication
  /api/v1/login:
    post:
      consumes:
      - application/json
      description: Login user with email and password
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api-gateway_internal_models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api-gateway_internal_models.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api-gateway_internal_models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api-gateway_internal_models.ErrorResponse'
      summary: User Login
      tags:
      - Authentication
  /api/v1/register:
    post:
      consumes:
      - application/json
      description: 'Register a new user with email, password, and personal information.
        Role field is optional and defaults to ''user'' if not provided. Valid roles:
        admin, user, manager, operator, analyst'
      parameters:
      - description: User registration data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api-gateway_internal_models.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/api-gateway_internal_models.RegisterResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api-gateway_internal_models.ErrorResponse'
      summary: User Registration
      tags:
      - Authentication
  /api/v1/reset-password:
    post:
      consumes:
      - application/json
      description: Reset password using token from email
      parameters:
      - description: Password reset data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api-gateway_internal_models.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api-gateway_internal_models.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api-gateway_internal_models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api-gateway_internal_models.ErrorResponse'
      summary: Reset Password
      tags:
      - Authentication
  /api/v1/users/{id}:
    delete:
      description: Delete user account. Users can only delete their own account. Admins
        can delete any user.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api-gateway_internal_models.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api-gateway_internal_models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api-gateway_internal_models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api-gateway_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete User
      tags:
      - Users
    get:
      description: Get user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api-gateway_internal_models.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api-gateway_internal_models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api-gateway_internal_models.ErrorResponse'
      summary: Get User
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: 'Update user information. All fields are optional. Role field must
        be one of: admin, user, manager, operator, analyst'
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: User update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api-gateway_internal_models.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api-gateway_internal_models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api-gateway_internal_models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api-gateway_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update User
      tags:
      - Users
  /api/v1/users/{id}/failed-attempts:
    get:
      description: Get failed login attempts for a specific user within a time window
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Start time in ISO format (e.g., 2023-01-01T00:00:00Z)
        in: query
        name: since
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api-gateway_internal_models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api-gateway_internal_models.ErrorResponse'
      summary: Get Failed Login Attempts
      tags:
      - Users
  /api/v1/users/{id}/login-history:
    get:
      description: Get paginated login history for a specific user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - default: 10
        description: Number of records to return
        in: query
        name: limit
        type: integer
      - default: 0
        description: Number of records to skip
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api-gateway_internal_models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api-gateway_internal_models.ErrorResponse'
      summary: Get User Login History
      tags:
      - Users
  /api/v1/users/{id}/password:
    put:
      consumes:
      - application/json
      description: Change user password
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Password change data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api-gateway_internal_models.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api-gateway_internal_models.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api-gateway_internal_models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api-gateway_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Change Password
      tags:
      - Users
  /api/v1/users/{id}/recent-logins:
    get:
      description: Get recent login history for a specific user (last 10 logins)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api-gateway_internal_models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api-gateway_internal_models.ErrorResponse'
      summary: Get Recent Login History
      tags:
      - Users
  /api/v1/users/me:
    get:
      description: Get current user's profile information from JWT token
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api-gateway_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get Current User Profile
      tags:
      - Authentication
  /api/v1/verify:
    get:
      description: Verify user email with token from registration email
      parameters:
      - description: Verification token
        in: query
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api-gateway_internal_models.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api-gateway_internal_models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api-gateway_internal_models.ErrorResponse'
      summary: Verify Email
      tags:
      - Authentication
swagger: "2.0"